Update range.metrics.R
Change default for single records to the width of a grid cell in the frame raster instead of arbitrary '1'. Same rule also for lat range that is less than the size of a pixel in the frame. raster, to avoid very small values - now framed to res of the map